Fixed paths to uImage in boot.{cmd,scr}

+++ mybuild.sh
@@ -1,20 +1,19 @@
 #!/bin/sh
 
-# @(#)mybuild.sh	1.11 03/21/16
-
 # ./OrangePI-Kernel/build/config.good/sun8iw7p1smp_lobo_defconfig.opiplus
 
-TMPF=`mktemp -d /tmp/XXXOPI_$$`
+TMPDIR=`mktemp -d /tmp/XXXOPI_$$`
 CURD=`pwd`
 OUTFOLDER=build
 SRCFOLDER=OrangePiPlus-Kernel
 LOBOSRCDIR=${SRCFOLDER}/OrangePI-Kernel
-SD_CARD_FILE=./sdcard.img
-ROOTPWD="toor" # Must be changed @first login
+SD_CARD_FILE=./nsdcard.img
+ROOTPWD="toor" # Must be changed @ first login
 
 end()
 {
-        sudo rm -f $TMPF
+	echo "Deleting $TMPDIR"
+        sudo rm -rf $TMPDIR
         exit 0
 }
 
@@ -22,14 +21,14 @@ trap end 9 2
 
 get_sources() {
 	if [ -d $SRCFOLDER ]; then
-		echo "Source folder ${SRCFOLDER} exists, updating ..."
+		echo "Git source folder ${SRCFOLDER} exists, updating ..."
 		(cd $SRCFOLDER ; git pull)
 	else
 		git clone https://github.com/j75/OrangePiPlus-Kernel.git
 	fi
 
 	if [ -d u-boot ]; then
-		echo "U-boot folder exists, updating ..."
+		echo "Git u-boot folder exists, updating ..."
 		(cd u-boot ; git pull)
 	else
 		git clone git://git.denx.de/u-boot.git
@@ -130,8 +129,6 @@ p
 1
 2048
 +100M
-t
-c
 n
 p
 
@@ -146,10 +143,24 @@ format() {
 	echo "Loop device is $LOOPDEV"
 	sudo kpartx -av $LOOPDEV
 	BASELDEV=`echo $LOOPDEV | cut -d '/' -f3`
-	mkfs.vfat -n BOOT  /dev/mapper/${BASELDEV}p1
+	mkfs.ext2 -L BOOT  /dev/mapper/${BASELDEV}p1
+	if [ $? -gt 0 ]; then
+		echo "Error formating /dev/mapper/${BASELDEV}p1"
+		sudo kpartx -dv $LOOPDEV
+		losetup -d $LOOPDEV
+		exit 7
+	fi
 	mkfs.ext4 -L Linux /dev/mapper/${BASELDEV}p2
+	FORMATP2=""
+	if [ $? -gt 0 ]; then
+		echo "Error formating /dev/mapper/${BASELDEV}p2"
+		FORMATP2="ko"
+	fi
 	sudo kpartx -dv $LOOPDEV
 	losetup -d $LOOPDEV
+	if [ -n $ FORMATP2 ]; then
+		exit 8
+	fi
 }
 
 copy_file() {
@@ -166,24 +177,32 @@ copy_boot_partition() {
 	echo "Loop device is $LOOPDEV"
 	sudo kpartx -av $LOOPDEV
 	BASELDEV=`echo $LOOPDEV | cut -d '/' -f3`
-	sudo mount /dev/mapper/${BASELDEV}p1 $TMPF
-	copy_file boot.scr $TMPF
+	sudo mount /dev/mapper/${BASELDEV}p1 $TMPDIR
+	if [ $? -gt 0 ]; then
+		echo "Error mounting /dev/mapper/${BASELDEV}p1 -> $TMPDIR"
+		sudo kpartx -dv $LOOPDEV
+		losetup -d $LOOPDEV
+		exit 1
+	fi
+	copy_file boot.scr $TMPDIR
 	rm boot.scr
-	#copy_file fex $TMPF
-	copy_file ${SRCFOLDER}/output/uImage $TMPF
+	#copy_file fex $TMPDIR
+	echo "Transferring uImage kernel to SD card boot partition"
+	ls -alF ${SRCFOLDER}/output/uImage*
+	sudo cp -a ${SRCFOLDER}/output/uImage* $TMPDIR
 	grep -v "^;" ${OUTFOLDER}/sys_config.fex | grep -v "^#" > script.fex
 	perl -pi -e 's|^max_freq.*|max_freq = 1100000000|g' script.fex
 	fex2bin script.fex script.bin
-	copy_file script.fex $TMPF
-	copy_file script.bin $TMPF
+	copy_file script.fex $TMPDIR
+	copy_file script.bin $TMPDIR
 	rm -f script.*
 	sudo sync
 	#
 	echo "Boot partition OK:"
-	ls -alF $TMPF
-	sudo umount $TMPF
+	ls -alF $TMPDIR
+	sudo umount $TMPDIR
 	if [ $? -gt 0 ]; then
-		echo "Error unmounting $TMPF"
+		echo "Error unmounting $TMPDIR"
 		exit 2
 	fi
 	sudo kpartx -dv $LOOPDEV
@@ -194,8 +213,8 @@ make_uboot_commands () {
 	echo "Creating boot.cmd file"
 	cat << EOF > boot.cmd
 setenv bootargs console=ttyS0 root=/dev/mmcblk0p1 rootwait panic=10 ${extra}
-ext2load mmc 0 0x43000000 boot/script.bin
-ext2load mmc 0 0x48000000 boot/uImage
+ext2load mmc 0 0x43000000 script.bin
+ext2load mmc 0 0x48000000 uImage
 bootm 0x48000000
 EOF
 	echo "Converting boot.cmd -> boot.scr"
@@ -207,42 +226,67 @@ copy_root_partition() {
 	echo "Loop device is $LOOPDEV"
 	sudo kpartx -av $LOOPDEV
 	BASELDEV=`echo $LOOPDEV | cut -d '/' -f3`
-	sudo mount /dev/mapper/${BASELDEV}p2 $TMPF
+	sudo mount /dev/mapper/${BASELDEV}p2 $TMPDIR
+	if [ $? -gt 0 ]; then
+		echo "Error mounting /dev/mapper/${BASELDEV}p2 -> $TMPDIR"
+		sudo kpartx -dv $LOOPDEV
+		losetup -d $LOOPDEV
+		exit 3
+	fi
 	#
-	sudo cp -r ${SRCFOLDER}/output/lib $TMPF
-	sudo mkdir ${TMPF}/boot
-	sudo cp ${SRCFOLDER}/output/*-* ${TMPF}/boot
+	echo "Transferring ${SRCFOLDER}/output/lib to SD card file"
+	sudo cp -r ${SRCFOLDER}/output/lib $TMPDIR
+	# rm build/source
+	sudo rm -f ${TMPDIR}/lib/modules/*/{build,source}
+	#
+	sudo mkdir ${TMPDIR}/boot
+	echo "Transferring kernel to SD card file root partition"
+	ls -alF ${SRCFOLDER}/output/*-*
+	sudo cp ${SRCFOLDER}/output/*-* ${TMPDIR}/boot
 	#
 	distro="jessie"
 	echo "Loading Debian $distro distribution"
-	sudo debootstrap --include=openssh-server,bash --arch=armhf --foreign $distro $TMPF
+	sudo debootstrap --include=openssh-server --arch=armhf --foreign $distro $TMPDIR
 	if [ $? -gt 0 ]; then
 		echo "Error creating Debian distribution"
 	else
-		echo "nameserver 127.0.1.1" | sudo tee ${TMPF}/etc/resolv.conf
-		sudo sh -c "echo 'orangepiplus' >  ${TMPF}/etc/hostname"
+		echo "nameserver 127.0.1.1" | sudo tee ${TMPDIR}/etc/resolv.conf
+		sudo sh -c "echo 'orangepiplus' >  ${TMPDIR}/etc/hostname"
+		#
 		cat <<EOT > /tmp/sources.list
 deb http://http.debian.net/debian $distro main contrib non-free
-deb-src http://http.debian.net/debian $distro main contrib non-free
+#deb-src http://http.debian.net/debian $distro main contrib non-free
+
 deb http://http.debian.net/debian $distro-updates main contrib non-free
-deb-src http://http.debian.net/debian $distro-updates main contrib non-free
+#deb-src http://http.debian.net/debian $distro-updates main contrib non-free
+
 deb http://security.debian.org/debian-security $distro/updates main contrib non-free
-deb-src http://security.debian.org/debian-security $distro/updates main contrib non-free
+#deb-src http://security.debian.org/debian-security $distro/updates main contrib non-free
 EOT
-		sudo cp -f /tmp/sources.list ${TMPF}/etc/apt/
+		sudo cp -f /tmp/sources.list ${TMPDIR}/etc/apt/
+		#
+		sudo rm -rf ${TMPDIR}/debootstrap
 		echo "Setting password $ROOTPWD -> should be in an ARM chrooted environment"
-		#sudo chroot ${TMPF} /bin/bash -c "(echo $ROOTPWD;echo $ROOTPWD;) | passwd root"
+		#sudo chroot ${TMPDIR} /bin/bash -c "(echo $ROOTPWD;echo $ROOTPWD;) | passwd root"
+		sudo sh -c "echo 'root:x:0:0:root:/root:/bin/bash' >  ${TMPDIR}/etc/passwd"
+		sudo sh -c "echo 'daemon:x:1:1:daemon:/usr/sbin:/usr/sbin/nologin' >>  ${TMPDIR}/etc/passwd"
+		sudo sh -c "echo 'bin:x:2:2:bin:/bin:/usr/sbin/nologin' >>  ${TMPDIR}/etc/passwd"
+		sudo sh -c "echo 'sys:x:3:3:sys:/dev:/usr/sbin/nologin' >>  ${TMPDIR}/etc/passwd"
+		sudo sh -c "echo 'nobody:x:65534:65534:nobody:/nonexistent:/usr/sbin/nologin' >>  ${TMPDIR}/etc/passwd"
+		sudo sh -c "echo 'root:\$6\$j3USIojv\$Vc1VrKa3j0MKXaLrEZyc2w3PnbRpBIPmt0ULvnIqquMPkWKW1fim4PBn.m0iC5BaZq609o27x3TIp8D0GKKBj/:16882:0:99999:7:::' >>  ${TMPDIR}/etc/shadow"
+		sudo chmod 640 ${TMPDIR}/etc/shadow
+		#
 		echo "Debian loaded"
 	fi
 	#
 	sudo sync
 	#
 	echo "Linux partition OK:"
-	ls -alF $TMPF
-	sudo umount $TMPF
+	ls -alF $TMPDIR
+	sudo umount $TMPDIR
 	if [ $? -gt 0 ]; then
-		echo "Error unmounting $TMPF"
-		exit 3
+		echo "Error unmounting $TMPDIR"
+		exit 4
 	fi
 	sudo kpartx -dv $LOOPDEV
 	losetup -d $LOOPDEV
@@ -271,7 +315,9 @@ check_exec () {
 	which $1 > /dev/null
 	if [ $? -gt 0 ]; then
 		echo "$1 not found ... exiting"
-		exit 4
+		exit 5
+	else
+		echo "  $1 exists ... good!"
 	fi
 }
 
@@ -279,7 +325,9 @@ check_group() {
 	id | grep $1 >/dev/null
 	if [ $? -gt 0 ]; then
 		echo "User does not belong to group $1 ... exiting"
-		exit 4
+		exit 6
+	else
+		echo "  user belongs to group $1 ... good!"
 	fi
 }
 
@@ -287,16 +335,17 @@ check_requirements() {
 	echo "Checking requirements for building..."
 	# kpartx mount.vfat losetup, dd,...
 	check_exec dd
+	check_exec chmod
 	check_exec losetup
 	check_exec kpartx
-	check_exec mkfs.vfat
+	check_exec mkfs.ext2
 	check_exec mkfs.ext4
 	check_exec debootstrap
 	#check_exec cdebootstrap
 	check_exec chroot
 	check_group disk
 	check_group sudo
-	echo "Good - all requirements are fullfiled!"
+	echo "All requirements are fullfiled!"
 }
 
 build_all () {
